requirements.txt
django>=3.0,<4.0
django-crispy-forms
crispy-forms-semantic-ui
django-registration-redux
openpyxl==3.0.3
numpy==1.20
numpy-financial==1.0.0
django_cron
cython
gunicorn
python-dateutil
reportlab==3.6.6
xhtml2pdf
psycopg2-binary